requirement.txt.txt
Appium-Python-Client==2.6.1;python_version >= '3.0'
